server/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-JWT-Extended==4.4.3
future==0.18.2
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
PyJWT==2.4.0
PyOpenGL==3.1.6
PyYAML==6.0
RPi.GPIO==0.7.1
serial==0.0.97
websockets==10.3
Werkzeug==2.2.2
